Fix get_all_intervals w.r.t. to expanding to inf
When expanding to +inf/-inf, we need consider if the input value is
between infinity and the first or last bound, or if it is exactly at the
first or last bound. If it equals the first or last bound, we should
receive multiple intervals. If it is not equal, we must only return an
interval expanding to infinity.